Add files via upload
The most important component of a website is having users. Real humans who can contribute to the website. If Facebook had no users then it would just be adverts. If blogs had no users then it would just be the ramblings of an author.
But in order to have users and associate data to user accounts, we need a way to register them and allow them to sign back into their accounts at a later date.
This means they will be giving us some information that we have to keep secure. This is what authentication is all about, in this project flask login/logout authentication created with password hashing and salting.